+++ app/src/main/java/com/techcourse/controller/annotation/RegisterController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.annotation;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
